Refactor: 서비스단 코드 분리
